refacto contactForm.js
